Move scrapy/command.py to scrapy/commands/base.py
