+++ scrapy/commands/crawl.py
@@ -1,5 +1,5 @@
 import os
-from scrapy.command import ScrapyCommand
+from scrapy.commands import ScrapyCommand
 from scrapy.utils.conf import arglist_to_dict
 from scrapy.exceptions import UsageError
