+++ scrapy/commands/edit.py
@@ -1,6 +1,6 @@
 import sys, os
 
-from scrapy.command import ScrapyCommand
+from scrapy.commands import ScrapyCommand
 from scrapy.exceptions import UsageError
 
 class Command(ScrapyCommand):
