+++ scrapy/commands/fetch.py
@@ -1,7 +1,7 @@
 from __future__ import print_function
 from w3lib.url import is_url
 
-from scrapy.command import ScrapyCommand
+from scrapy.commands import ScrapyCommand
 from scrapy.http import Request
 from scrapy.exceptions import UsageError
 from scrapy.utils.spider import spidercls_for_request, DefaultSpider
