+++ scrapy/commands/shell.py
@@ -6,7 +6,7 @@ See documentation in docs/topics/shell.rst
 
 from threading import Thread
 
-from scrapy.command import ScrapyCommand
+from scrapy.commands import ScrapyCommand
 from scrapy.shell import Shell
 from scrapy.http import Request
 from scrapy.utils.spider import spidercls_for_request, DefaultSpider
