+++ scrapy/commands/view.py
@@ -1,5 +1,4 @@
-from scrapy.command import ScrapyCommand
-from scrapy.commands import fetch
+from scrapy.commands import fetch, ScrapyCommand
 from scrapy.utils.response import open_in_browser
 
 class Command(fetch.Command):
